requirements.txt
fastapi
uvicorn
google-generativeai
edge-tts
python-dotenv
jinja2
python-multipart
